requirements.txt
streamlit==1.33.0
pillow==10.3.0
groq==0.9.0
python-dotenv==0.21.0
PyPDF2==3.0.1
google-cloud-documentai==0.2.0
